top-comment-words/requirements.txt
kafka-python
redis
redis-server
nltk
